+++ projects/services-map-state/src/lib/types/map-state.ts
@@ -16,6 +16,7 @@ export interface IProjFitOptions {
   fitToProjectionExtent?: boolean
   fitToBbox?: TGeoExtent
   fitToNativeBbox?: TGeoExtent
+  viewSetExtent?: boolean // default true
 }
 
 export interface IMapState {